Rename PositionResponsibility classes
