+++ src/main/java/dev/fmagallanes97/backendportfolio/positionresponsibility/PosResponsibility.java
@@ -10,7 +10,7 @@ import javax.persistence.*;
 @Getter
 @Setter
 @Table(name = "position_responsibility")
-public class PositionResponsibility {
+public class PosResponsibility {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
