+++ src/main/java/dev/fmagallanes97/backendportfolio/positionresponsibility/PosResponsibilityRepository.java
@@ -1,0 +1,9 @@
+package dev.fmagallanes97.backendportfolio.positionresponsibility;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface PosResponsibilityRepository extends JpaRepository<PosResponsibility, Long> {
+    public List<PosResponsibility> findAllByPositionId(Long id);
+}
