+++ src/main/java/dev/fmagallanes97/backendportfolio/positionresponsibility/dto/PosResponsibilityRequest.java
@@ -3,7 +3,7 @@ package dev.fmagallanes97.backendportfolio.positionresponsibility.dto;
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.Size;
 
-public record PositionResponsibilityRequest(
+public record PosResponsibilityRequest(
         @Size(min = 4, max = 255, message = "this value must be between 4 and 255 characters")
         @NotEmpty(message = "this attribute is mandatory")
         String description
