+++ src/main/java/dev/fmagallanes97/backendportfolio/positionresponsibility/dto/PosResponsibilityResponse.java
@@ -1,6 +1,6 @@
 package dev.fmagallanes97.backendportfolio.positionresponsibility.dto;
 
-public record PositionResponsibilityResponse(
+public record PosResponsibilityResponse(
         Long id,
         String description
 ) {
